backend/requirements.txt
Django>=3.1.3
djangorestframework>=3.12.2
djangorestframework-jwt>=1.11.0
requests>=2.25.0
pytest>=6.2.1
freezegun>=1.0.0
pytz>=2020.4
PyJWT>=1.7.1